blocklength in writeappendeddata should be integer type
[[Imported from SVN: r3897]]

+++ grid/io/file/vtk/vtkwriter.hh
@@ -1213,7 +1213,7 @@ namespace Dune
       SimpleStream stream(s);
 
       // write length before each data block
-      float blocklength;
+      unsigned int blocklength;
 
       // point data
       for (FunctionIterator it=vertexdata.begin(); it!=vertexdata.end(); ++it)
